requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
tqdm>=4.66.4
pyyaml>=6.0.1
